Naji 17 05 refactoring

+++ src/main/java/com/boefcity/projectmanagement/controller/ProjectController.java
@@ -1,6 +1,6 @@
 package com.boefcity.projectmanagement.controller;
 
-import com.boefcity.projectmanagement.config.SessionUtility;
+import com.boefcity.projectmanagement.config.AppUtility;
 import com.boefcity.projectmanagement.model.*;
 import com.boefcity.projectmanagement.service.ProjectService;
 import com.boefcity.projectmanagement.service.UserService;
@@ -30,81 +30,79 @@ public class ProjectController {
     public String projectsDisplay(HttpSession session, Model model,
                                   RedirectAttributes redirectAttributes) {
 
-        if (SessionUtility.isNotAuthenticated(session, redirectAttributes)) {
+        if (AppUtility.isNotAuthenticated(session, redirectAttributes)) {
             return "redirect:/users/loginDisplay";
         }
 
-        Long userId = (Long) session.getAttribute("userId");
-        User user = userService.findUserById(userId);
+        Long currentUserId = (Long) session.getAttribute("userId");
+        User currentUser = userService.findUserById(currentUserId);
 
-        Role role = user.getUserRole();
-
-        if (Role.ADMIN.equals(role) || Role.MANAGER.equals(role)) {
-            List<Project> projectList = projectService.findAllProjects();
-            model.addAttribute("projectList", projectList);
-            return "/project/projectList";
+        List<Project> projectList;
+        if (AppUtility.isAdminOrManager(currentUser)) {
+            projectList = projectService.findAllProjects();
         } else {
-            List<Project> projectList = user.getProjects();
-            model.addAttribute("projectList", projectList);
-            return "/project/projectList";
+            projectList = currentUser.getProjects();
         }
+
+        model.addAttribute("projectList", projectList);
+        return "/project/projectList";
     }
+
     @GetMapping("/addForm")
     public String projectsAddFormDisplay(Model model, HttpSession session,
                                          RedirectAttributes redirectAttributes) {
 
-        if (SessionUtility.isNotAuthenticated(session, redirectAttributes)) {
+        if (AppUtility.isNotAuthenticated(session, redirectAttributes)) {
             return "redirect:/users/loginDisplay";
         }
 
-        Long userId = (Long) session.getAttribute("userId");
-        User user = userService.findUserById(userId);
+        Long currentUserId = (Long) session.getAttribute("userId");
+        User currentUser = userService.findUserById(currentUserId);
 
-        Role role = user.getUserRole();
-
-        if (Role.ADMIN.equals(role) || Role.MANAGER.equals(role)) {
-            model.addAttribute("project", new Project());
-            return "/project/projectAddForm";
-        } else {
-            redirectAttributes.addFlashAttribute("message", "User not authorized to add new projects");
+        if (!AppUtility.isAdminOrManager(currentUser)) {
+            redirectAttributes.addFlashAttribute("message", "User not authorized to add new projects.");
             return "redirect:/projects/display";
         }
+
+        model.addAttribute("project", new Project());
+        return "/project/projectAddForm";
     }
+
 
     @PostMapping("/create")
     public String createProject(@ModelAttribute Project project,
                                 HttpSession session,
                                 RedirectAttributes redirectAttributes) {
 
-        if (SessionUtility.isNotAuthenticated(session, redirectAttributes)) {
+        if (AppUtility.isNotAuthenticated(session, redirectAttributes)) {
             return "redirect:/users/loginDisplay";
         }
 
-        Long userId = (Long) session.getAttribute("userId");
-        User user = userService.findUserById(userId);
+        Long currentUserId = (Long) session.getAttribute("userId");
+        User currentUser = userService.findUserById(currentUserId);
 
-        Role role = user.getUserRole();
-
-        if (Role.ADMIN.equals(role) || Role.MANAGER.equals(role)) {
-            try {
-                projectService.createProject(project);
-                redirectAttributes.addFlashAttribute("message",
-                        "You have successfully created a new project: " + project.getProjectName());
-                return "redirect:/projects/display";
-            } catch (Exception e) {
-                redirectAttributes.addFlashAttribute("message", "Something went wrong. Please try again.");
-            }
-        } else {
-            redirectAttributes.addFlashAttribute("message", "User not authorized to create a new project");
+        if (!AppUtility.isAdminOrManager(currentUser)) {
+            redirectAttributes.addFlashAttribute("message", "User not authorized to create a new project.");
+            return "redirect:/projects/display";
         }
+
+        try {
+            projectService.createProject(project);
+            redirectAttributes.addFlashAttribute("message",
+                    "You have successfully created a new project: " + project.getProjectName());
+        } catch (Exception e) {
+            redirectAttributes.addFlashAttribute("message", "Something went wrong. Please try again.");
+        }
+
         return "redirect:/projects/display";
     }
 
-// OVERVIEW DISPLAY
+
+    // OVERVIEW DISPLAY
     @GetMapping("/overviewDisplay")
     public String projectsOverviewDisplay(HttpSession session, Model model, @RequestParam Long projectId,
                                     RedirectAttributes redirectAttributes) {
-        if (SessionUtility.isNotAuthenticated(session, redirectAttributes)) {
+        if (AppUtility.isNotAuthenticated(session, redirectAttributes)) {
             return "redirect:/users/loginDisplay";
         }
 
@@ -136,14 +134,19 @@ public class ProjectController {
     public String assignUserToProject(@RequestParam Long projectId, @RequestParam Long userId,
                                       HttpSession session, RedirectAttributes redirectAttributes) {
 
-        if (SessionUtility.isNotAuthenticated(session, redirectAttributes)) {
+        if (AppUtility.isNotAuthenticated(session, redirectAttributes)) {
             return "redirect:/users/loginDisplay";
+        }
+
+        if (projectId == null || userId == null) {
+            redirectAttributes.addFlashAttribute("message", "User or project information is missing. Please try again.");
+            return "redirect:/projects/display";
         }
 
         Long currentUserId = (Long) session.getAttribute("userId");
         User currentUser = userService.findUserById(currentUserId);
 
-        if (!isAdminOrManager(currentUser)) {
+        if (!AppUtility.isAdminOrManager(currentUser)) {
             redirectAttributes.addFlashAttribute("message", "User not authorized to assign users to this project.");
             return "redirect:/projects/overviewDisplay?projectId=" + projectId;
         }
@@ -158,35 +161,42 @@ public class ProjectController {
         } catch (RuntimeException e) {
             redirectAttributes.addFlashAttribute("message", "Something went wrong. User was not assigned.");
         }
+
         return "redirect:/projects/overviewDisplay?projectId=" + projectId;
     }
+
 
     @PostMapping("/delete/{projectToDeleteId}")
     public String deleteProject(@PathVariable Long projectToDeleteId, HttpSession session,
                                 RedirectAttributes redirectAttributes) {
 
-        if (SessionUtility.isNotAuthenticated(session, redirectAttributes)) {
+        if (AppUtility.isNotAuthenticated(session, redirectAttributes)) {
             return "redirect:/projects/display";
         }
 
         Long currentUserId = (Long) session.getAttribute("userId");
         User currentUser = userService.findUserById(currentUserId);
 
-        Role role = currentUser.getUserRole();
-        if (!Role.ADMIN.equals(role) && !Role.MANAGER.equals(role)) {
-            return "errorPage";
+        if (!AppUtility.isAdminOrManager(currentUser)) {
+            redirectAttributes.addFlashAttribute("message", "User not authorized to delete projects.");
+            return "redirect:/projects/display";
         }
 
-        Project projectToDelete = projectService.findProjectById(projectToDeleteId);
-        if (projectToDelete != null) {
-            projectService.deleteProjectById(projectToDeleteId);
-            redirectAttributes.addFlashAttribute("message", projectToDelete.getProjectName() + " was successfully deleted.");
-        } else {
-            redirectAttributes.addFlashAttribute("message", "Project not found.");
+        try {
+            Project projectToDelete = projectService.findProjectById(projectToDeleteId);
+            if (projectToDelete == null) {
+                redirectAttributes.addFlashAttribute("message", "Project not found.");
+            } else {
+                projectService.deleteProjectById(projectToDeleteId);
+                redirectAttributes.addFlashAttribute("message", projectToDelete.getProjectName() + " was successfully deleted.");
+            }
+        } catch (RuntimeException e) {
+            redirectAttributes.addFlashAttribute("message", "Something went wrong. The project was not deleted.");
         }
 
         return "redirect:/projects/display";
     }
+
 
     @PostMapping("/removeUser")
     public String removeUserFromProject(@RequestParam Long userId,
@@ -194,16 +204,8 @@ public class ProjectController {
                                         HttpSession session,
                                         RedirectAttributes redirectAttributes) {
 
-        if (SessionUtility.isNotAuthenticated(session, redirectAttributes)) {
+        if (AppUtility.isNotAuthenticated(session, redirectAttributes)) {
             return "redirect:/projects/display";
-        }
-
-        Long currentUserId = (Long) session.getAttribute("userId");
-        User currentUser = userService.findUserById(currentUserId);
-        Role role = currentUser.getUserRole();
-        if (!Role.ADMIN.equals(role) && !Role.MANAGER.equals(role)) {
-            redirectAttributes.addFlashAttribute("message", "You are not authorized to remove assigned users");
-            return "redirect:/projects/overviewDisplay?projectId=" + projectId;
         }
 
         if (projectId == null || userId == null) {
@@ -211,14 +213,22 @@ public class ProjectController {
             return "redirect:/projects/display";
         }
 
+        Long currentUserId = (Long) session.getAttribute("userId");
+        User currentUser = userService.findUserById(currentUserId);
+
+        if (!AppUtility.isAdminOrManager(currentUser)) {
+            redirectAttributes.addFlashAttribute("message", "You are not authorized to remove assigned users.");
+            return "redirect:/projects/overviewDisplay?projectId=" + projectId;
+        }
+
         try {
             projectService.removeUserFromProject(userId, projectId);
             redirectAttributes.addFlashAttribute("message", "User was successfully removed from the project.");
-            return "redirect:/projects/overviewDisplay?projectId=" + projectId;
         } catch (Exception e) {
-            redirectAttributes.addFlashAttribute("message", "Failed to remove user from the project due to unexpected error.");
-            return "redirect:/errorPage";
+            redirectAttributes.addFlashAttribute("message", "Failed to remove user from the project due to an unexpected error.");
         }
+
+        return "redirect:/projects/overviewDisplay?projectId=" + projectId;
     }
 
 
@@ -228,22 +238,21 @@ public class ProjectController {
                                          HttpSession session,
                                          RedirectAttributes redirectAttributes) {
 
-        if (SessionUtility.isNotAuthenticated(session, redirectAttributes)) {
+        if (AppUtility.isNotAuthenticated(session, redirectAttributes)) {
             return "redirect:/users/loginDisplay";
         }
 
         Long currentUserId = (Long) session.getAttribute("userId");
         User currentUser = userService.findUserById(currentUserId);
-        Role role = currentUser.getUserRole();
 
-        if (!Role.ADMIN.equals(role) && !Role.MANAGER.equals(role)) {
-            redirectAttributes.addFlashAttribute("message", "User not authorized to edit this project");
+        if (!AppUtility.isAdminOrManager(currentUser)) {
+            redirectAttributes.addFlashAttribute("message", "User not authorized to edit this project.");
             return "redirect:/projects/overviewDisplay?projectId=" + projectId;
         }
 
         Project projectToEdit = projectService.findProjectById(projectId);
         if (projectToEdit == null) {
-            redirectAttributes.addFlashAttribute("message", "Project to edit was not found");
+            redirectAttributes.addFlashAttribute("message", "Project to edit was not found.");
             return "redirect:/projects/overviewDisplay?projectId=" + projectId;
         }
 
@@ -251,41 +260,35 @@ public class ProjectController {
         return "/project/editForm";
     }
 
+
     @PostMapping("/editForm/{projectId}")
     public String editProject(@PathVariable Long projectId,
-                                @ModelAttribute("project") Project projectDetails,
-                                HttpSession session,
-                                RedirectAttributes redirectAttributes) {
+                              @ModelAttribute("project") Project projectDetails,
+                              HttpSession session,
+                              RedirectAttributes redirectAttributes) {
 
-        if (SessionUtility.isNotAuthenticated(session, redirectAttributes)) {
+        if (AppUtility.isNotAuthenticated(session, redirectAttributes)) {
             return "redirect:/users/loginDisplay";
         }
 
         Long currentUserId = (Long) session.getAttribute("userId");
         User currentUser = userService.findUserById(currentUserId);
-        Role role = currentUser.getUserRole();
 
-        if (!Role.ADMIN.equals(role) && !Role.MANAGER.equals(role)) {
-            redirectAttributes.addFlashAttribute("message", "User not authorized to edit this project");
+        if (!AppUtility.isAdminOrManager(currentUser)) {
+            redirectAttributes.addFlashAttribute("message", "User not authorized to edit this project.");
             return "redirect:/projects/overviewDisplay?projectId=" + projectId;
         }
 
         Project existingProject = projectService.findProjectById(projectId);
         if (existingProject == null) {
-            redirectAttributes.addFlashAttribute("message", "Project to edit was not found");
+            redirectAttributes.addFlashAttribute("message", "Project to edit was not found.");
             return "redirect:/projects/overviewDisplay?projectId=" + projectId;
         }
 
-
         projectService.editProject(projectId, projectDetails);
-
-        redirectAttributes.addFlashAttribute("message", "Project updated successfully");
+        redirectAttributes.addFlashAttribute("message", "Project updated successfully.");
         return "redirect:/projects/overviewDisplay?projectId=" + projectId;
     }
 
-    private boolean isAdminOrManager(User user) {
-        Role role = user.getUserRole();
-        return Role.ADMIN.equals(role) || Role.MANAGER.equals(role);
-    }
 
 }
